Polish help section: nested navigation, contact/FAQ/feedback screens, form improvements
